Verify serialport availability before opening
On Windows, nrfjprog will hang when opening J-Link devices that are in a
bad state. A device may get into this state if left connected to the PC
overnight.

nRF Connect performed this check as part of the old serial port selector
implementation, but it fits better here now that we have the nrfjprog
interaction in this library.