Remove default sdReq for SoftDevice init packet
I am not sure why 0xFE was given as a default sdReq for the SoftDevice
init packet. It makes the DFU fail when trying to DFU
s132_softdevice_3.0.0 and connectivity application.

I think it makes most sense to leave the sdReq for the SoftDevice init
packet empty if it is not explicitly provided. When leaving it empty,
the DFU seems to work fine regardless of what SoftDevice is present on
the device from before.